requirements.txt
alembic==1.4.2
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
vk-api==11.8.0
wcwidth==0.1.9
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
